+++ doc/javadoc/package-search-index.js
@@ -1,0 +1,1 @@
+packageSearchIndex = [{"l":"All Packages","url":"allpackages-index.html"},{"l":"app"},{"l":"app.controller"},{"l":"backend"},{"l":"client"},{"l":"protocol"},{"l":"server"}]