+++ src/main/java/app/Runner.java
@@ -1,0 +1,10 @@
+package app;
+
+/**
+ * Running the app.
+ */
+public class Runner {
+  public static void main(String[] args) {
+    Main.main(args);
+  }
+}