+++ src/main/java/app/controller/SingleplayerController.java
@@ -1,0 +1,12 @@
+package app.controller;
+
+import javafx.stage.Stage;
+
+class SingleplayerController extends Controller {
+
+  private static final String FXMLPATH = "view/singleplayer.fxml";
+
+  SingleplayerController(Stage stage) {
+    super(stage, FXMLPATH);
+  }
+}